Remove popularity logic from sql.py; that's specifically for the loader workflow and doens't belong there. Implement recomputation of percentiles from expired file cache.
