+++ src/cc_catalog_airflow/dags/util/loader/sql.py
@@ -235,32 +235,6 @@ def upsert_records_to_image_table(
     postgres.run(upsert_query)
 
 
-def _build_popularity_query(popularity_fields):
-    """
-    Given a list of fields used in popularity data calculations, build a query
-    returning all rows with at least one popularity metric.
-    """
-    # SELECT predicate for each popularity field
-    selection_qs = []
-    # WHERE predicate excluding null values for each field
-    field_not_null_qs = []
-    for field in popularity_fields:
-        selection_qs.append(f"meta_data->>'{field}' AS {field}")
-        field_not_null_qs.append(f"meta_data->>'{field}' IS NOT NULL")
-    selections = ', '.join(selection_qs)sql.py
-    field_not_null = ' OR '.join(field_not_null_qs)
-    return f"SELECT identifier, provider, {selections} FROM image " \
-           f"WHERE ({field_not_null})"
-
-
-def dump_selection_to_tsv(postgres_conn_id, query, tsv_file_name):
-    postgres = PostgresHook(postgres_conn_id=postgres_conn_id)
-    postgres.copy_expert(
-        f"COPY ({query}) TO '{tsv_file_name}' "
-        f"WITH CSV HEADER DELIMITER E'\t'"
-    )
-
-
 def drop_load_table(postgres_conn_id, identifier):
     load_table = _get_load_table_name(identifier)
     postgres = PostgresHook(postgres_conn_id=postgres_conn_id)
